+++ app/src/main/java/miridih/command/CommandInvoker.java
@@ -1,4 +1,4 @@
-package miridih.commands;
+package miridih.command;
 
 import java.util.Stack;
