+++ app/src/main/java/miridih/command/UndoCommand.java
@@ -1,4 +1,4 @@
-package miridih.commands;
+package miridih.command;
 
 import miridih.model.CanvasModel;
